ProxyAuthenticationPolicy.Factory -> ProxyAuthenticationPolicyFactory
